refactor: move logic from vaadin-menu-bar to the mixin
