+++ packages/menu-bar/src/vaadin-menu-bar.d.ts
@@ -6,55 +6,9 @@
 import { DisabledMixin } from '@vaadin/a11y-base/src/disabled-mixin.js';
 import { ElementMixin } from '@vaadin/component-base/src/element-mixin.js';
 import { ThemableMixin } from '@vaadin/vaadin-themable-mixin/vaadin-themable-mixin.js';
-import { MenuBarMixin } from './vaadin-menu-bar-mixin.js';
+import { type MenuBarItem, MenuBarMixin } from './vaadin-menu-bar-mixin.js';
 
-export interface MenuBarItem {
-  /**
-   * Text to be set as the menu button component's textContent.
-   */
-  text?: string;
-  /**
-   * Text to be set as the menu button's tooltip.
-   * Requires a `<vaadin-tooltip slot="tooltip">` element to be added inside the `<vaadin-menu-bar>`.
-   */
-  tooltip?: string;
-  /**
-   * The component to represent the button content.
-   * Either a tagName or an element instance. Defaults to "vaadin-menu-bar-item".
-   */
-  component?: HTMLElement | string;
-  /**
-   * If true, the button is disabled and cannot be activated.
-   */
-  disabled?: boolean;
-  /**
-   * Theme(s) to be set as the theme attribute of the button, overriding any theme set on the menu bar.
-   */
-  theme?: string[] | string;
-  /**
-   * Array of submenu items.
-   */
-  children?: SubMenuItem[];
-
-  /**
-   * Class/classes to be set to the class attribute of the button.
-   */
-  className?: string;
-}
-
-export interface SubMenuItem {
-  text?: string;
-  component?: HTMLElement | string;
-  disabled?: boolean;
-  theme?: string[] | string;
-  checked?: boolean;
-  className?: string;
-  children?: SubMenuItem[];
-}
-
-export interface MenuBarI18n {
-  moreOptions: string;
-}
+export { MenuBarItem, MenuBarI18n, SubMenuItem } from './vaadin-menu-bar-mixin.js';
 
 /**
  * Fired when a submenu item or menu bar button without children is clicked.
@@ -114,67 +68,6 @@ export interface MenuBarEventMap extends HTMLElementEventMap, MenuBarCustomEvent
  * @fires {CustomEvent} item-selected - Fired when a submenu item or menu bar button without children is clicked.
  */
 declare class MenuBar extends MenuBarMixin(DisabledMixin(ElementMixin(ThemableMixin(HTMLElement)))) {
-  /**
-   * Defines a hierarchical structure, where root level items represent menu bar buttons,
-   * and `children` property configures a submenu with items to be opened below
-   * the button on click, Enter, Space, Up and Down arrow keys.
-   *
-   * #### Example
-   *
-   * ```js
-   * menubar.items = [
-   *   {
-   *     text: 'File',
-   *     className: 'file',
-   *     children: [
-   *       {text: 'Open', className: 'file open'}
-   *       {text: 'Auto Save', checked: true},
-   *     ]
-   *   },
-   *   {component: 'hr'},
-   *   {
-   *     text: 'Edit',
-   *     children: [
-   *       {text: 'Undo', disabled: true},
-   *       {text: 'Redo'}
-   *     ]
-   *   },
-   *   {text: 'Help'}
-   * ];
-   * ```
-   */
-  items: MenuBarItem[];
-
-  /**
-   * The object used to localize this component.
-   * To change the default localization, replace the entire
-   * `i18n` object with a custom one.
-   *
-   * To update individual properties, extend the existing i18n object like so:
-   * ```
-   * menuBar.i18n = {
-   *   ...menuBar.i18n,
-   *   moreOptions: 'More options'
-   * }
-   * ```
-   *
-   * The object has the following JSON structure and default values:
-   * ```
-   * {
-   *   moreOptions: 'More options'
-   * }
-   * ```
-   */
-  i18n: MenuBarI18n;
-
-  /**
-   * A space-delimited list of CSS class names
-   * to set on each sub-menu overlay element.
-   *
-   * @attr {string} overlay-class
-   */
-  overlayClass: string;
-
   addEventListener<K extends keyof MenuBarEventMap>(
     type: K,
     listener: (this: MenuBar, ev: MenuBarEventMap[K]) => void,
@@ -186,12 +79,6 @@ declare class MenuBar extends MenuBarMixin(DisabledMixin(ElementMixin(ThemableMi
     listener: (this: MenuBar, ev: MenuBarEventMap[K]) => void,
     options?: EventListenerOptions | boolean,
   ): void;
-
-  /**
-   * A callback for the `_theme` property observer.
-   * It propagates the host theme to the buttons and the sub menu.
-   */
-  protected _themeChanged(theme: string | null): void;
 }
 
 declare global {
